requirements.txt
-i https://pypi.org/simple
cachetools==4.1.0; python_version ~= '3.5'
certifi==2020.6.20
chardet==3.0.4
click==7.1.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
flask==1.1.2
google-auth-oauthlib==0.4.1
google-auth==1.18.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
greenlet==0.4.16
gspread==3.6.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
itsdangerous==1.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
jinja2==2.11.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
markupsafe==1.1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
msgpack==1.0.0
neovim==0.3.1
oauth2client==4.1.3
oauthlib==3.1.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1-modules==0.2.8
pyasn1==0.4.8
pynvim==0.4.1
pytelegrambotapi==3.7.1
python-dotenv==0.13.0
requests-oauthlib==1.3.0
requests==2.24.0
rsa==4.6; python_version >= '3'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==1.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
